Enable seq to be imported from baker
Enable the seq() function to be imported from the baker module like
it is in the recipe module so that it can be used as baker.seq() if
preferred or to avoid namespace issues.

The current import can be left in place for backwards compatibility
or to continue importing seq as before.

Closes #75