Add test models and test to replicate the problem
The test will fail with v1.3.3 (it passes with v1.3.2) like:

  AttributeError: 'ManyToManyRel' object has no attribute 'has_default'

The problem seems to be that I'm using a ManyToManyField with a specific
through table, which I am then making an instance of, passing both its
foreignkey fields with the dunder syntax. Passing instances for any of
these directly instead will make the test pass. Removing the
ManyToManyField will also make the test pass.